Refactor: add a currentCat variable in model.
A current data holder in the code is a very useful thought. It's a black box so you don't have to think out how to get the currentCat when you need one. The whole thinking is very much related to that of a state machine.
The original code also made the mistake of copying model inside the octopus, and letting the view directly modify the copied version.